+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -2,6 +2,7 @@ define([
   "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/units.js",
   "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/unit_groups.js",
 ], function (gwoUnit, gwoGroup) {
+  const vehicleFactories = [gwoUnit.vehicleFactory, gwoUnit.vehicleFactoryAdvanced]
   return {
     cards: [
       { id: "gwaio_enable_bot_aa" },
@@ -423,6 +424,7 @@ define([
       { id: "gwaio_damage_titans", units: gwoGroup.titans },
       { id: "gwaio_speed_titans", units: gwoGroup.titans },
       { id: "gwaio_combat_titans", units: gwoGroup.titans },
+      { id: "gwaio_cooldown_vehicles", units: vehicleFactories },
     ],
   };
 });