+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -224,6 +224,7 @@ function gwoCard() {
       ];
       const expandedCards = [
         "gwaio_combat_titans",
+        "gwaio_cooldown_vehicles",
         "gwaio_damage_titans",
         "gwaio_enable_planetaryradar",
         "gwaio_health_titans",